tidy: Fix "multiplication implicit widening" in libmythtv (1).
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

1) Cast multiplication arguments to a larger type.

2) Convert a variable to the appropriate type.

3) Convert a couple of #define to typed static constexpr.

+++ mythtv/libs/libmythtv/captions/cc708window.cpp
@@ -171,7 +171,7 @@ void CC708Window::Resize(uint new_rows, uint new_columns)
     if (new_rows < m_true_row_count || new_columns < m_true_column_count)
     {
       delete [] m_text;
-      m_text = new CC708Character [new_rows * new_columns];
+      m_text = new CC708Character [static_cast<size_t>(new_rows) * new_columns];
       m_true_row_count = new_rows;
       m_true_column_count = new_columns;
       m_pen.m_row = 0;
@@ -194,7 +194,7 @@ void CC708Window::Resize(uint new_rows, uint new_columns)
 
         // Expand the array if the new size exceeds the current capacity
         // in either dimension.
-        auto *new_text = new CC708Character[new_rows * new_columns];
+        auto *new_text = new CC708Character[static_cast<size_t>(new_rows) * new_columns];
         m_pen.m_column = 0;
         m_pen.m_row = 0;
         uint i = 0;
